🚚 move: Separate document and entity packages
