+++ src/main/java/com/mulmeong/feed/api/dto/in/FeedHashtagUpdateDto.java
@@ -1,7 +1,7 @@
 package com.mulmeong.feed.api.dto.in;
 
 import com.mulmeong.feed.api.domain.entity.Feed;
-import com.mulmeong.feed.api.domain.entity.FeedHashtag;
+import com.mulmeong.feed.api.domain.document.FeedHashtag;
 import com.mulmeong.feed.api.domain.event.FeedHashtagUpdateEvent;
 import com.mulmeong.feed.api.domain.model.Hashtag;
 import com.mulmeong.feed.api.vo.in.FeedHashtagUpdateVo;
